ml/requirements.txt
torch>=2.0.0
torchvision
transformers>=4.36.0
diffusers>=0.28.0
pillow
numpy
accelerate
